requirements.txt
Django==5.2
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
django-redis==5.4.0
redis==6.0.0